% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivue-package.R
\docType{package}
\name{ivue-package}
\alias{ivue}
\alias{ivue-package}
\title{ivue: Interactive 3D Visualization of Data and Graphs}
\description{
A package for reusable visualization of numerical data and embedded graphs.
This development version contains the package scaffold; plotting functions
have not yet been migrated from gflow.
}
\author{
\strong{Maintainer}: Pawel Gajer \email{pgajer@gmail.com}

Authors:
\itemize{
  \item Pawel Gajer \email{pgajer@gmail.com}
}

}
\keyword{internal}
